+++ parsley-core/src/ghc/Parsley/Internal/Common/QueueLike.hs
@@ -1,0 +1,9 @@
+module Parsley.Internal.Common.QueueLike (module Parsley.Internal.Common.QueueLike) where
+
+class QueueLike q where
+  empty :: q a
+  null :: q a -> Bool
+  size :: q a -> Int
+  enqueue :: a -> q a -> q a
+  dequeue :: q a -> (a, q a)
+  enqueueAll :: [a] -> q a -> q a